reorganize common service modules
