+++ common-data/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
     api(project(":common-api"))
     api(project(":infra-api"))
     api("com.fasterxml.jackson.core:jackson-databind")
+    api("com.squareup.okhttp3:okhttp")
     api("com.google.dagger:dagger")
     api("io.undertow:undertow-core")
     api("javax.inject:javax.inject")
