+++ common-data/src/main/java/org/example/age/common/service/data/AvsLocation.java
@@ -1,23 +1,23 @@
-package org.example.age.common.service.config;
+package org.example.age.common.service.data;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.net.HostAndPort;
 import okhttp3.HttpUrl;
+import org.example.age.data.crypto.SecureId;
 import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
 
 /**
- * URL location of a site.
+ * URL location of the age verification service.
  *
- * <p>A real implementation would used HTTPS.</p>
+ * <p>A real implementation would use HTTPS.</p>
  */
 @Value.Immutable
 @DataStyle
-@JsonSerialize(as = ImmutableSiteLocation.class)
-@JsonDeserialize(as = ImmutableSiteLocation.class)
-public interface SiteLocation {
+@JsonSerialize(as = ImmutableAvsLocation.class)
+@JsonDeserialize(as = ImmutableAvsLocation.class)
+public interface AvsLocation {
 
     /** Creates a builder for the location. */
     static Builder builder(String host, int port) {
@@ -29,41 +29,41 @@ public interface SiteLocation {
         return new Builder().hostAndPort(hostAndPort);
     }
 
-    /** Host and port of the site. */
+    /** Host and port of the age verification service. */
     HostAndPort hostAndPort();
 
-    /** Path of the API to process an age certificate. */
+    /** Path of the API to create a verification session. */
     @Value.Default
-    default String ageCertificatePath() {
-        return "api/age-certificate";
+    default String verificationSessionPath() {
+        return "api/verification-session";
     }
 
     /** Path to redirect users to in order to continue age verification. */
     String redirectPath();
 
-    /** URL of the API to process an age certificate. */
+    /** URL of the API to create a verification session. */
     @Value.Derived
-    @JsonIgnore
-    default HttpUrl ageCertificateUrl() {
+    default HttpUrl verificationSessionUrl(String siteId) {
         return new HttpUrl.Builder()
                 .scheme("http")
                 .host(hostAndPort().getHost())
                 .port(hostAndPort().getPort())
-                .addPathSegments(ageCertificatePath())
+                .addPathSegments(verificationSessionPath())
+                .addQueryParameter("site-id", siteId)
                 .build();
     }
 
     /** URL to redirect users to in order to continue age verification. */
     @Value.Derived
-    @JsonIgnore
-    default HttpUrl redirectUrl() {
+    default HttpUrl redirectUrl(SecureId requestId) {
         return new HttpUrl.Builder()
                 .scheme("http")
                 .host(hostAndPort().getHost())
                 .port(hostAndPort().getPort())
                 .addPathSegments(redirectPath())
+                .addQueryParameter("request-id", requestId.toString())
                 .build();
     }
 
-    final class Builder extends ImmutableSiteLocation.Builder {}
+    final class Builder extends ImmutableAvsLocation.Builder {}
 }
