+++ common/src/main/java/org/example/age/common/avs/verification/internal/VerificationManagerImpl.java
@@ -15,8 +15,8 @@ import org.example.age.common.api.data.AuthMatchDataExtractor;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
 import org.example.age.common.avs.store.RegisteredSiteConfigStore;
 import org.example.age.common.avs.store.VerifiedUserStore;
-import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.crypto.internal.AuthMatchDataEncryptor;
+import org.example.age.common.service.data.SiteLocation;
 import org.example.age.common.service.store.PendingStore;
 import org.example.age.common.service.store.PendingStoreFactory;
 import org.example.age.data.certificate.AgeCertificate;
