+++ common/src/test/java/org/example/age/common/avs/verification/VerificationManagerTest.java
@@ -25,8 +25,8 @@ import org.example.age.common.avs.store.VerifiedUserStore;
 import org.example.age.common.avs.verification.internal.Verification;
 import org.example.age.common.avs.verification.internal.VerificationManager;
 import org.example.age.common.avs.verification.internal.VerificationManagerModule;
-import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
+import org.example.age.common.service.data.SiteLocation;
 import org.example.age.common.service.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.data.crypto.SecureId;
