+++ site-service/src/main/java/org/example/age/site/service/SiteService.java
@@ -11,7 +11,7 @@ import org.example.age.api.Dispatcher;
 import org.example.age.api.JsonSender;
 import org.example.age.api.StatusCodeSender;
 import org.example.age.common.api.data.AuthMatchData;
-import org.example.age.common.service.config.AvsLocation;
+import org.example.age.common.service.data.AvsLocation;
 import org.example.age.data.certificate.SignedAgeCertificate;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.infra.service.client.RequestDispatcher;
