+++ site-service/src/main/java/org/example/age/site/service/config/SiteConfig.java
@@ -2,7 +2,7 @@ package org.example.age.site.service.config;
 
 import java.security.PublicKey;
 import java.time.Duration;
-import org.example.age.common.service.config.AvsLocation;
+import org.example.age.common.service.data.AvsLocation;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
