+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsService.java
@@ -15,8 +15,8 @@ import org.example.age.api.JsonSender;
 import org.example.age.api.StatusCodeSender;
 import org.example.age.avs.api.AvsApi;
 import org.example.age.common.api.data.AuthMatchData;
-import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.crypto.internal.AuthMatchDataEncryptor;
+import org.example.age.common.service.data.SiteLocation;
 import org.example.age.data.certificate.AgeCertificate;
 import org.example.age.data.certificate.SignedAgeCertificate;
 import org.example.age.data.certificate.VerificationRequest;
