+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsServiceModule.java
@@ -5,9 +5,9 @@ import dagger.Module;
 import java.security.PrivateKey;
 import org.example.age.avs.api.AvsApi;
 import org.example.age.avs.api.AvsApiModule;
-import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.crypto.internal.AuthMatchDataEncryptorModule;
 import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
+import org.example.age.common.service.data.SiteLocation;
 import org.example.age.common.service.data.internal.DataMapperModule;
 import org.example.age.infra.service.client.RequestDispatcherModule;
 import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
